requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.17.1
python-decouple==3.8
sqlparse==0.5.5
psycopg[binary]
dj-database-url
beautifulsoup4
lxml
setuptools
selenium
undetected-chromedriver
confluent-kafka
gunicorn
whitenoise
certifi
django-redis
redis
pytest
pytest-django
pytest-cov
responses
fakeredis
pre-commit
ruff
